refactor: organize scripts into subpackages
