+++ docs/MIGRATION_GUIDE.md
@@ -4,9 +4,9 @@ The following thin wrapper modules were relocated to improve project organizatio
 
 | Old Path | New Path |
 | --- | --- |
-| `advanced_qubo_optimization.py` | `scripts/advanced_qubo_optimization.py` |
+| `advanced_qubo_optimization.py` | `scripts/entrypoints/advanced_qubo_optimization.py` |
 | `simplified_quantum_integration_orchestrator.py` | `scripts/session/simplified_quantum_integration_orchestrator.py` |
-| `unified_database_management_system.py` | `scripts/unified_database_management_system.py` |
+| `unified_database_management_system.py` | `scripts/entrypoints/unified_database_management_system.py` |
 | `unified_disaster_recovery_system.py` | `scripts/unified_disaster_recovery_system.py` |
 | `unified_legacy_cleanup_system.py` | `scripts/unified_legacy_cleanup_wrapper.py` |
 | `unified_monitoring_optimization_system.py` | `scripts/unified_monitoring_optimization_system.py` |
