+++ documentation/COMPLETE_TECHNICAL_SPECIFICATIONS_WHITEPAPER.md
@@ -42,7 +42,7 @@ The gh_COPILOT Toolkit v4.0 represents a revolutionary enterprise-grade automati
 | `unified_monitoring_optimization_system.py` | Present |
 | `unified_script_generation_system.py` | Present |
 | `unified_session_management_system.py` | Present |
-| `unified_database_management_system.py` | Present |
+| `scripts/entrypoints/unified_database_management_system.py` | Present |
 | `unified_disaster_recovery_system.py` | Planned – partial implementation |
 | `unified_legacy_cleanup_system.py` | Present in `scripts/` directory |
 | `web_gui_integration_system.py` | Planned – partial integration |
