+++ documentation/ENTERPRISE_SCRIPT_CONSOLIDATION_STRATEGY.md
@@ -20,11 +20,11 @@
 - **Action:** COMPLETE - No further action needed
 
 ### **2. ✅ DATABASE MANAGEMENT** (COMPLETED)
-**Status:** Successfully consolidated into `scripts/unified_database_management_system.py`
+**Status:** Successfully consolidated into `scripts/entrypoints/unified_database_management_system.py`
 - **Migration ID:** DB_MIGRATION_20250707_074327
 - **Scripts Archived:** 9 database management scripts
 - **Archive Location:** `scripts/archived_database_scripts/DB_MIGRATION_20250707_074327/`
-- **Canonical Version:** `scripts/unified_database_management_system.py`
+- **Canonical Version:** `scripts/entrypoints/unified_database_management_system.py`
 - **Action:** COMPLETE - All database management functionality unified
 
 ### **3. 🚨 DISASTER RECOVERY** (REQUIRES CONSOLIDATION)
