+++ scripts/automation/quantum_integration_orchestrator.py
@@ -20,7 +20,7 @@ from typing import List, Tuple
 
 from tqdm import tqdm
 
-from scripts.advanced_qubo_optimization import solve_qubo_bruteforce
+from scripts.entrypoints.advanced_qubo_optimization import solve_qubo_bruteforce
 from scripts.validation.secondary_copilot_validator import SecondaryCopilotValidator
 from quantum.quantum_database_search import quantum_search_sql
 from quantum.ibm_backend import init_ibm_backend
@@ -99,7 +99,7 @@ class EnterpriseUtility:
         self.logger.info(f"{TEXT_INDICATORS['info']} Running QUBO demo")
 
         try:
-            from advanced_qubo_optimization import (
+            from scripts.entrypoints.advanced_qubo_optimization import (
                 EnterpriseUtility as QuboUtil,
             )
         except ImportError as exc:  # pragma: no cover - import guard
